+++ src/main/java/com/agilecheckup/persistency/entity/EmployeeAssessmentScore.java
@@ -4,7 +4,10 @@ import com.agilecheckup.persistency.entity.score.PillarScore;
 import com.agilecheckup.persistency.entity.score.Scorable;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBAttribute;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBDocument;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
 import java.util.Map;
@@ -21,6 +24,6 @@ public class EmployeeAssessmentScore implements Scorable {
   private Map<String, PillarScore> pillarIdToPillarScoreMap;
 
   @DynamoDBAttribute
-  private Integer score;
+  private Double score;
 
 }
