+++ src/main/java/com/agilecheckup/persistency/entity/score/Scorable.java
@@ -2,5 +2,5 @@ package com.agilecheckup.persistency.entity.score;
 
 public interface Scorable {
 
-  public Integer getScore();
+  public Double getScore();
 }
